// src/rundown-execution/application/enums/rundown-event-type.ts
export enum RundownEventType {
  ACTIVATED = 'ACTIVATED',
  REHEARSE = 'REHEARSE',
  DEACTIVATED = 'DEACTIVATED',
  RESET = 'RESET',
  TAKEN = 'TAKEN',
  SET_NEXT = 'SET_NEXT',
  PART_INSERTED_AS_ON_AIR = 'PART_INSERTED_AS_ON_AIR',
  PART_INSERTED_AS_NEXT = 'PART_INSERTED_AS_NEXT',
  PIECE_INSERTED = 'PIECE_INSERTED',
  PIECE_REPLACED = 'PIECE_REPLACED',
  PIECE_STOPPED = 'PIECE_STOPPED',
  INFINITE_PIECES_UPDATED = 'INFINITE_PIECES_UPDATED',
  RUNDOWN_CREATED = 'RUNDOWN_CREATED',
  RUNDOWN_UPDATED = 'RUNDOWN_UPDATED',
  RUNDOWN_DELETED = 'RUNDOWN_DELETED',
  SEGMENT_CREATED = 'SEGMENT_CREATED',
  SEGMENT_UPDATED = 'SEGMENT_UPDATED',
  SEGMENT_DELETED = 'SEGMENT_DELETED',
  SEGMENT_UNSYNCED = 'SEGMENT_UNSYNCED',
  PART_CREATED = 'PART_CREATED',
  PART_UPDATED = 'PART_UPDATED',
  PART_DELETED = 'PART_DELETED',
  PART_UNSYNCED = 'PART_UNSYNCED'
}